Split ShadowRenderer into class for local and clsas for directional lights (#4915)
