+++ src/scene/renderer/shadow-map-cache.js
@@ -11,22 +11,22 @@ import { ShadowMap } from './shadow-map.js';
 class ShadowMapCache {
     constructor() {
         // maps a shadow map key to an array of shadow maps in the cache
-        this.shadowMapCache = new Map();
+        this.cache = new Map();
     }
 
     destroy() {
         this.clear();
-        this.shadowMapCache = null;
+        this.cache = null;
     }
 
     // remove all shadowmaps from the cache
     clear() {
-        this.shadowMapCache.forEach((shadowMaps) => {
+        this.cache.forEach((shadowMaps) => {
             shadowMaps.forEach((shadowMap) => {
                 shadowMap.destroy();
             });
         });
-        this.shadowMapCache.clear();
+        this.cache.clear();
     }
 
     // generates a string key for the shadow map required by the light
@@ -42,7 +42,7 @@ class ShadowMapCache {
 
         // get matching shadow buffer from the cache
         const key = this.getKey(light);
-        const shadowMaps = this.shadowMapCache.get(key);
+        const shadowMaps = this.cache.get(key);
         if (shadowMaps && shadowMaps.length) {
             return shadowMaps.pop();
         }
@@ -56,11 +56,11 @@ class ShadowMapCache {
     // returns shadow map for the light back to the cache
     add(light, shadowMap) {
         const key = this.getKey(light);
-        const shadowMaps = this.shadowMapCache.get(key);
+        const shadowMaps = this.cache.get(key);
         if (shadowMaps) {
             shadowMaps.push(shadowMap);
         } else {
-            this.shadowMapCache.set(key, [shadowMap]);
+            this.cache.set(key, [shadowMap]);
         }
     }
 }
